+++ rollup.config.js
@@ -4,26 +4,27 @@ import sucrase from '@rollup/plugin-sucrase';
 
 import workerLoader from 'rollup-plugin-web-worker-loader';
 import glslify from 'rollup-plugin-glslify';
+import copy from 'rollup-plugin-copy';
 
 import pkg from './package.json';
 const external = [
   ...Object.keys(pkg.peerDependencies),
   ...Object.keys(pkg.dependencies),
-  ...Object.keys(pkg.devDependencies),
+  ...Object.keys(pkg.devDependencies)
 ];
 
-const config = (test) => {
+const config = test => {
   if (!test) {
     return {
       input: 'src/index.js',
       output: { file: 'dist/index.js', format: 'esm' }
-    }
+    };
   }
   return {
     input: `tests/${test}/index.spec.js`,
     output: { format: 'cjs' }
-  }
-}
+  };
+};
 
 export default {
   ...config(process.env.TEST),
@@ -33,13 +34,16 @@ export default {
     glslify(),
     workerLoader({
       targetPlatform: 'browser',
-      inline: true,
+      inline: true
     }),
     sucrase({
-      exclude: "node_modules/*",
+      exclude: 'node_modules/*',
       transforms: ['jsx'],
-      production: true,
+      production: true
     }),
     commonjs(),
+    copy({
+      targets: [{ src: 'src/index.d.ts', dest: 'dist' }]
+    })
   ]
-}
+};