+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
 - Issue templates for Github.
-- Add `transparentColor` and `useTransparentColor` to allow the deck.gl to set a color to be "transparent."
+- Add `transparentColor` and `useTransparentColor` to allow the layer to set a color to be "transparent" (or use the automatically calculated one when colormaps are set).
 
 ### Changed