+++ CHANGELOG.md
@@ -5,9 +5,10 @@
 ### Added
 
 - Issue templates for Github.
-- Add `transparentColor` and `useTransparentColor` to allow the layer to set a color to be "transparent" (or use the automatically calculated one when colormaps are set).
+- Add `transparentColor` to allow the layer to set a color to be "transparent" (or use the automatically calculated one when colormaps are set).
 
 ### Changed
+
 - Allow for `tileLayer` zoom level to be offset by model matrix scaling.
 
 ## 0.8.2